+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## Feature
   - เพิ่มการรับ Request จาก Admin
   - แก้ไข SQL ให้รองรับหลาย vendor
-  - ส่งค่าวันที่ API Server เพื่อการตรวจสอบเวลา
+  - ส่งค่าวันที่ API Server 
+  - เพิ่มการรับ Request จาก admin
 
 # 3.8.10
 ## Feature